Fix script for step build

+++ build-pi-packages.sh
@@ -3,27 +3,17 @@
 set -x
 set -e
 ROOTDIR=$(pwd)
-PKG_DIR=$ROOTDIR/packages
 OCAML_VERSION=4.14.2
 
-cd $ROOTDIR
+# Check root user
+if [ "$EUID" -ne 0 ]; then
+    echo "❌ERROR: You must run script with root privileges!"
+    exit 1
+fi
+
 # Build vyos-1.x for RPI4
-echo "Clone VyOS-1.x code"
-cd $ROOTDIR/vyos-build/scripts/package-build/vyos-1x/
-find . -maxdepth 1 ! -name "package.toml" ! -name "build.py" -exec rm -rf {} \;
-git clone --recurse-submodules https://github.com/vyos/vyos-1x
-
-# Patch vyos-1.x code for RPI console ttyAMA0
-echo "Patch code Vyos-1.x for RPI4"
-cd ./vyos-1x 
-patch -p1 < $ROOTDIR/patches/vyos-1.x/vyos-1.x-rpi4-patches.diff
-git add .
-git config --global user.name "TCNGUYEN" && git config --global user.email "trancaonguyendn@gmail.com"
-git commit -m "Fix code for RPI"
-cd ../
-
-# Build Vyos-1.x packages
-echo "Build Vyos-1.x"
+cd ${ROOTDIR}/vyos-build/scripts/package-build/vyos-1x/
+echo "Building Vyos-1.x"
 rm -rf /usr/lib/libvyosconfig.so.0
 if [ ! -f /.dockerenv ]; then
     sysctl -w net.ipv4.conf.lo.forwarding=1
@@ -32,9 +22,11 @@ fi
 ./build.py
 find . -maxdepth 1 -type f -name "*.deb" | grep -E "^(./)?(libvyosconfig0_|vyos-1x_)" | xargs cp -t $ROOTDIR/packages
 
-cd $ROOTDIR
 # Build telegraf for RPI4
-cd ./vyos-build/scripts/package-build/telegraf
+cd ${ROOTDIR}/vyos-build/scripts/package-build/telegraf
 rm -rf telegraf
 ./build.py
 find . -maxdepth 1 -type f -name "*.deb" | grep -E "^(./)?(telegraf_)" | xargs cp -t $ROOTDIR/packages
+
+# Return to ROOTDIR
+cd ${ROOTDIR}
